Move recur_map to sardanautils (avoid duplication of code)
The same implementation of recur_map is placed in two different modules.
Avoid duplication of code.
