+++ src/sardana/macroserver/msmacromanager.py
@@ -58,7 +58,7 @@ from taurus.core.util.codecs import CodecFactory
 from sardana.sardanadefs import ElementType
 from sardana.sardanamodulemanager import ModuleManager
 from sardana.sardanaexception import format_exception_only_str
-from sardana.sardanautils import is_pure_str, is_non_str_seq
+from sardana.sardanautils import is_pure_str, is_non_str_seq, recur_map
 
 from sardana.macroserver.msmanager import MacroServerManager
 from sardana.macroserver.msmetamacro import MACRO_TEMPLATE, MacroLibrary, \
@@ -139,22 +139,6 @@ def is_macro(macro, abs_file=None, logger=None):
     else:
         return False
     return True
-
-
-def recur_map(fun, data, keep_none=False):
-    """Recursive map. Similar to map, but maintains the list objects structure
-
-    :param fun: <callable> the same purpose as in map function
-    :param data: <object> the same purpose as in map function
-    :param keep_none: <bool> keep None elements without applying fun
-    """
-    if hasattr(data, "__iter__") and not isinstance(data, str):
-        return [recur_map(fun, elem, keep_none) for elem in data]
-    else:
-        if keep_none is True and data is None:
-            return data
-        else:
-            return fun(data)
 
 
 def is_flat_list(obj):
